django_questions/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.2.0
packaging==24.2
propcache==0.3.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
yarl==1.18.3
